Improved interface to microphysics. Support non-physical species.
For Maestro, read the species names, A, and Z from the `job_info` file
and if the species cannot be identified as an element, then add it to
the network anyway using the A and Z in `job_info`. Otherwise use A
and Z from the periodic table lookup. This allows the use of
descriptive species names like "ash."